CVSL-1454 PR comments addressed
