+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/jobs/TimeOutLicencesService.kt
@@ -4,17 +4,11 @@ import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.entity.AuditEvent
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.entity.CrdLicence
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.entity.LicenceEvent
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.repository.AuditEventRepository
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.repository.LicenceEventRepository
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.repository.LicenceRepository
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.NotifyService
+import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.LicenceService
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.ReleaseDateService
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.WorkingDaysService
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.util.AuditEventType
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.util.LicenceEventType
 import java.time.Clock
 import java.time.LocalDate
 
@@ -23,11 +17,9 @@ class TimeOutLicencesService(
   @Value("\${hardstop.enabled}") private val hardStopEnabled: Boolean,
   private val licenceRepository: LicenceRepository,
   private val releaseDateService: ReleaseDateService,
-  private val auditEventRepository: AuditEventRepository,
-  private val licenceEventRepository: LicenceEventRepository,
   private val workingDaysService: WorkingDaysService,
   private val clock: Clock,
-  private val notifyService: NotifyService,
+  private val licenceService: LicenceService,
 ) {
 
   companion object {
@@ -38,7 +30,7 @@ class TimeOutLicencesService(
   fun timeOutLicences() {
     log.info("Job to runTimeOutLicencesService started")
     val jobExecutionDate = LocalDate.now(clock)
-    if (workingDaysService.isNonWorkingDay(jobExecutionDate)) {
+    if (workingDaysService.isNonWorkingDay(jobExecutionDate) || !hardStopEnabled) {
       return
     }
     val licencesToTimeOut = licenceRepository.getAllLicencesToTimeOut().filter {
@@ -53,49 +45,7 @@ class TimeOutLicencesService(
     log.info("TimeOutLicencesServiceJob updated status TIMED_OUT on ${licencesToTimeOut.size} licences")
   }
 
-  private fun updateLicencesStatus(licences: List<CrdLicence>, reason: String? = null) {
-    val timeOutLicences = licences.map { it.timeOut() }
-
-    licenceRepository.saveAllAndFlush(timeOutLicences)
-
-    timeOutLicences.map { licence ->
-      auditEventRepository.saveAndFlush(
-        AuditEvent(
-          licenceId = licence.id,
-          username = "SYSTEM",
-          fullName = "SYSTEM",
-          eventType = AuditEventType.SYSTEM_EVENT,
-          summary = "${reason ?: "Licence automatically timed out"} for ${licence.forename} ${licence.surname}",
-          detail = "ID ${licence.id} type ${licence.typeCode} status ${licence.statusCode.name} version ${licence.version}",
-        ),
-      )
-
-      licenceEventRepository.saveAndFlush(
-        LicenceEvent(
-          licenceId = licence.id,
-          eventType = LicenceEventType.TIMED_OUT,
-          username = "SYSTEM",
-          forenames = "SYSTEM",
-          surname = "SYSTEM",
-          eventDescription = "${reason ?: "Licence automatically timed out"} for ${licence.forename} ${licence.surname}",
-        ),
-      )
-
-      // previously approved now edited licence that is timed out
-      if (hardStopEnabled && licence.versionOfId != null) {
-        val com = licence.responsibleCom
-        if (com != null) {
-          notifyService.sendEditedLicenceTimedOutEmail(
-            com.email,
-            "${com.firstName} ${com.lastName}",
-            licence.forename!!,
-            licence.surname!!,
-            licence.crn,
-            licence.licenceStartDate,
-            licence.id.toString(),
-          )
-        }
-      }
-    }
+  private fun updateLicencesStatus(licences: List<CrdLicence>) {
+    licences.map { licence -> licenceService.timeout(licence, "due to reaching hard stop") }
   }
 }
