+++ src/test/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/jobs/TimeOutLicencesServiceTest.kt
@@ -1,31 +1,17 @@
 package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service
 
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.mockito.ArgumentCaptor
-import org.mockito.kotlin.any
-import org.mockito.kotlin.anyOrNull
-import org.mockito.kotlin.argumentCaptor
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.reset
 import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
-import org.mockito.kotlin.verifyNoInteractions
 import org.mockito.kotlin.whenever
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.context.SecurityContext
 import org.springframework.security.core.context.SecurityContextHolder
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.entity.AuditEvent
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.entity.CrdLicence
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.entity.LicenceEvent
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.repository.AuditEventRepository
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.repository.LicenceEventRepository
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.repository.LicenceRepository
 import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.jobs.TimeOutLicencesService
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.util.AuditEventType
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.util.LicenceEventType
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.util.LicenceStatus
 import java.time.Clock
 import java.time.Instant
 import java.time.LocalDate
@@ -34,20 +20,16 @@ import java.time.ZoneId
 class TimeOutLicencesServiceTest {
   private val licenceRepository = mock<LicenceRepository>()
   private val releaseDateService = mock<ReleaseDateService>()
-  private val auditEventRepository = mock<AuditEventRepository>()
-  private val licenceEventRepository = mock<LicenceEventRepository>()
   private val workingDaysService = mock<WorkingDaysService>()
-  private val notifyService = mock<NotifyService>()
+  private val licenceService = mock<LicenceService>()
 
   private val service = TimeOutLicencesService(
     hardStopEnabled = true,
     licenceRepository,
     releaseDateService,
-    auditEventRepository,
-    licenceEventRepository,
     workingDaysService,
     clock,
-    notifyService,
+    licenceService,
   )
 
   @BeforeEach
@@ -62,9 +44,8 @@ class TimeOutLicencesServiceTest {
     reset(
       licenceRepository,
       releaseDateService,
-      auditEventRepository,
-      licenceEventRepository,
       workingDaysService,
+      licenceService,
     )
   }
 
@@ -88,10 +69,6 @@ class TimeOutLicencesServiceTest {
     verify(licenceRepository, times(1)).getAllLicencesToTimeOut()
 
     verify(licenceRepository, times(0)).saveAllAndFlush(emptyList())
-
-    verify(auditEventRepository, times(0)).saveAndFlush(any())
-    verify(licenceEventRepository, times(0)).saveAndFlush(any())
-    verifyNoInteractions(notifyService)
   }
 
   @Test
@@ -108,47 +85,9 @@ class TimeOutLicencesServiceTest {
 
     service.timeOutLicences()
 
-    val licenceCaptor = argumentCaptor<List<CrdLicence>>()
-    val auditCaptor = ArgumentCaptor.forClass(AuditEvent::class.java)
-    val eventCaptor = ArgumentCaptor.forClass(LicenceEvent::class.java)
-
     verify(licenceRepository, times(1)).getAllLicencesToTimeOut()
 
-    verify(licenceRepository, times(1)).saveAllAndFlush(licenceCaptor.capture())
-
-    assertThat(licenceCaptor.firstValue[0])
-      .extracting("statusCode", "updatedByUsername")
-      .isEqualTo(listOf(LicenceStatus.TIMED_OUT, "SYSTEM"))
-
-    verify(auditEventRepository, times(1)).saveAndFlush(auditCaptor.capture())
-    verify(licenceEventRepository, times(1)).saveAndFlush(eventCaptor.capture())
-    verifyNoInteractions(notifyService)
-
-    assertThat(auditCaptor.value)
-      .extracting("licenceId", "username", "fullName", "eventType", "summary", "detail")
-      .isEqualTo(
-        listOf(
-          1L,
-          "SYSTEM",
-          "SYSTEM",
-          AuditEventType.SYSTEM_EVENT,
-          "Licence automatically timed out for ${aLicenceEntity.forename} ${aLicenceEntity.surname}",
-          "ID ${aLicenceEntity.id} type ${aLicenceEntity.typeCode} status ${LicenceStatus.TIMED_OUT} version ${aLicenceEntity.version}",
-        ),
-      )
-
-    assertThat(eventCaptor.value)
-      .extracting("licenceId", "eventType", "username", "forenames", "surname", "eventDescription")
-      .isEqualTo(
-        listOf(
-          1L,
-          LicenceEventType.TIMED_OUT,
-          "SYSTEM",
-          "SYSTEM",
-          "SYSTEM",
-          "Licence automatically timed out for ${aLicenceEntity.forename} ${aLicenceEntity.surname}",
-        ),
-      )
+    verify(licenceService, times(1)).timeout(aLicenceEntity, "due to reaching hard stop")
   }
 
   @Test
@@ -174,119 +113,10 @@ class TimeOutLicencesServiceTest {
 
     service.timeOutLicences()
 
-    val licenceCaptor = argumentCaptor<List<CrdLicence>>()
-    val auditCaptor = ArgumentCaptor.forClass(AuditEvent::class.java)
-    val eventCaptor = ArgumentCaptor.forClass(LicenceEvent::class.java)
-
     verify(licenceRepository, times(1)).getAllLicencesToTimeOut()
 
-    verify(licenceRepository, times(1)).saveAllAndFlush(licenceCaptor.capture())
-
-    verifyNoInteractions(notifyService)
-
-    assertThat(licenceCaptor.allValues.size).isEqualTo(1)
-
-    assertThat(licenceCaptor.firstValue[0])
-      .extracting("id", "statusCode", "updatedByUsername")
-      .isEqualTo(listOf(1L, LicenceStatus.TIMED_OUT, "SYSTEM"))
-
-    verify(auditEventRepository, times(1)).saveAndFlush(auditCaptor.capture())
-    verify(licenceEventRepository, times(1)).saveAndFlush(eventCaptor.capture())
-
-    assertThat(auditCaptor.value)
-      .extracting("licenceId", "username", "fullName", "eventType", "summary", "detail")
-      .isEqualTo(
-        listOf(
-          1L,
-          "SYSTEM",
-          "SYSTEM",
-          AuditEventType.SYSTEM_EVENT,
-          "Licence automatically timed out for ${aLicenceEntity.forename} ${aLicenceEntity.surname}",
-          "ID ${aLicenceEntity.id} type ${aLicenceEntity.typeCode} status ${LicenceStatus.TIMED_OUT} version ${aLicenceEntity.version}",
-        ),
-      )
-
-    assertThat(eventCaptor.value)
-      .extracting("licenceId", "eventType", "username", "forenames", "surname", "eventDescription")
-      .isEqualTo(
-        listOf(
-          1L,
-          LicenceEventType.TIMED_OUT,
-          "SYSTEM",
-          "SYSTEM",
-          "SYSTEM",
-          "Licence automatically timed out for ${aLicenceEntity.forename} ${aLicenceEntity.surname}",
-        ),
-      )
-  }
-
-  @Test
-  fun `should send notification when a previously approved now edited licence is timed out`() {
-    val approvedThenEditedLicence = aLicenceEntity.copy(
-      id = 2L,
-      versionOfId = 1L,
-    )
-    whenever(workingDaysService.isNonWorkingDay(LocalDate.now(clock))).thenReturn(false)
-    whenever(licenceRepository.getAllLicencesToTimeOut()).thenReturn(
-      listOf(
-        approvedThenEditedLicence,
-      ),
-    )
-    whenever(releaseDateService.isInHardStopPeriod(any(), anyOrNull())).thenReturn(
-      true,
-    )
-
-    service.timeOutLicences()
-
-    val licenceCaptor = argumentCaptor<List<CrdLicence>>()
-    val auditCaptor = ArgumentCaptor.forClass(AuditEvent::class.java)
-    val eventCaptor = ArgumentCaptor.forClass(LicenceEvent::class.java)
-
-    verify(licenceRepository, times(1)).getAllLicencesToTimeOut()
-
-    verify(licenceRepository, times(1)).saveAllAndFlush(licenceCaptor.capture())
-
-    assertThat(licenceCaptor.firstValue[0])
-      .extracting("statusCode", "updatedByUsername")
-      .isEqualTo(listOf(LicenceStatus.TIMED_OUT, "SYSTEM"))
-
-    verify(auditEventRepository, times(1)).saveAndFlush(auditCaptor.capture())
-    verify(licenceEventRepository, times(1)).saveAndFlush(eventCaptor.capture())
-    verify(notifyService, times(1)).sendEditedLicenceTimedOutEmail(
-      approvedThenEditedLicence.responsibleCom?.email,
-      "${approvedThenEditedLicence.responsibleCom?.firstName} ${approvedThenEditedLicence.responsibleCom?.lastName}",
-      approvedThenEditedLicence.forename!!,
-      approvedThenEditedLicence.surname!!,
-      approvedThenEditedLicence.crn!!,
-      approvedThenEditedLicence.licenceStartDate,
-      approvedThenEditedLicence.id.toString(),
-    )
-
-    assertThat(auditCaptor.value)
-      .extracting("licenceId", "username", "fullName", "eventType", "summary", "detail")
-      .isEqualTo(
-        listOf(
-          2L,
-          "SYSTEM",
-          "SYSTEM",
-          AuditEventType.SYSTEM_EVENT,
-          "Licence automatically timed out for ${approvedThenEditedLicence.forename} ${approvedThenEditedLicence.surname}",
-          "ID ${approvedThenEditedLicence.id} type ${approvedThenEditedLicence.typeCode} status ${LicenceStatus.TIMED_OUT} version ${approvedThenEditedLicence.version}",
-        ),
-      )
-
-    assertThat(eventCaptor.value)
-      .extracting("licenceId", "eventType", "username", "forenames", "surname", "eventDescription")
-      .isEqualTo(
-        listOf(
-          2L,
-          LicenceEventType.TIMED_OUT,
-          "SYSTEM",
-          "SYSTEM",
-          "SYSTEM",
-          "Licence automatically timed out for ${approvedThenEditedLicence.forename} ${approvedThenEditedLicence.surname}",
-        ),
-      )
+    verify(licenceService, times(1)).timeout(aLicenceEntity, "due to reaching hard stop")
+    verify(licenceService, times(0)).timeout(anIneligibleLicence, "due to reaching hard stop")
   }
 
   private companion object {
